Fix `ensureContentScript` too
