+++ src/background/deployment.ts
@@ -27,7 +27,7 @@ import { selectExtensions } from "@/store/extensionsSelectors";
 import { RegistryId, UUID } from "@/core";
 import { maybeGetLinkedApiClient } from "@/services/apiClient";
 import { queueReactivateTab } from "@/contentScript/messenger/api";
-import { forEachTab } from "./util";
+import { forEachTab } from "@/chrome";
 import { parse as parseSemVer, satisfies, SemVer } from "semver";
 import { ExtensionOptionsState } from "@/store/extensionsTypes";
 import extensionsSlice from "@/store/extensionsSlice";
