+++ src/background/initBrowserCommands.ts
@@ -20,7 +20,7 @@ import { Tab } from "@/mv3/api";
 import { Target } from "@/types";
 import { expectContext } from "@/utils/expectContext";
 import { canReceiveContentScript } from "@/utils/permissions";
-import { ensureContentScript } from "./util";
+import { ensureContentScript } from "./contentScript";
 
 async function handleCommand(command: string, tab: Tab): Promise<void> {
   if (command !== "toggle-quick-bar" || !canReceiveContentScript(tab.url)) {
