+++ src/background/navigation.ts
@@ -16,7 +16,7 @@
  */
 
 import { reactivateTab, handleNavigate } from "@/contentScript/messenger/api";
-import { forEachTab } from "@/background/util";
+import { forEachTab } from "@/chrome";
 import { canAccessTab } from "webext-tools";
 import { Target } from "@/types";
